Cookie Banner Updated.
The cookie banner is now the same for every device. Mobile and Desktop are now using the same cookie banner and selector instead of separate ones for every device.
The JavaScript has been adapted accordingly and now does not contain the script for the elements on mobile devices.
All these changes are leading to a overall performance boost and a faster and visual better design.

Affected files:
- assests/global/cookie-banner.js
- discord/authentication.html
- discord/callback.html
- discord/index.html
- download/index.html
- download/v1.0.html
- download/v1.1.html
- download/v2.0.html
- download/v3.0.html
- download/v4.0.html
- download/v4.1.html
- download/v4.2.html
- download/v5.0.html
- helpcenter/debug-menu.html
- helpcenter/faq.html
- helpcenter/import-world.html
- helpcenter/index.html
- index.html
- legal/accessibility-statement.html
- legal/advertisement-terms.html
- legal/application-guidelines.html
- legal/eula.html
- legal/github-privacy.html
- legal/imprint.html
- legal/index.html
- legal/privacy.html
- legal/terms.html
- legal/use-our-world.html
- linkout.html
- login/index.html
- login/return.html
- v5-0.html